VSCE: Autoinstall ato
- refactor ui
